Extract shared benchmark test fixtures and entry-bar helper
